requirements.txt
flask-smorest
python-dotenv
flask_restful
#sqlalchamy
#flask-sqlalchamy
flask-jwt-extended
passlib
flask-migrate